+++ packages/ccui/ui/tree/src/tree.tsx
@@ -1,7 +1,7 @@
 import type { CSSProperties, VNode } from 'vue'
 import type { FlattenedTreeNode, TreeDropInfo, TreeDropPosition, TreeNodeKey, TreeProps } from './tree-types'
 import { computed, defineComponent, h, nextTick, onUnmounted, ref, shallowRef, toRef, watch } from 'vue'
-import { CaretRightOutlined } from '@vue3-ccui/icons'
+import { CaretRightOutlined } from '@vaebe/ccui-icons'
 import { useNamespace } from '../../shared/hooks/use-namespace'
 import { useVirtualList } from '../../shared/hooks/use-virtual-list'
 import { computeNextCheckedKeys, useCheckedDerived } from './composables/use-tree-check'