+++ packages/docs/.vitepress/theme/index.ts
@@ -1,7 +1,7 @@
 import type { EnhanceAppContext } from 'vitepress'
 import DemoPreview, { useComponents } from '@vitepress-code-preview/container'
 import DefaultTheme from 'vitepress/theme'
-import ccui from 'vue3-ccui/ui/vue-ccui'
+import ccui from '@vaebe/ccui/ui/vue-ccui'
 import './styles/index.css'
 import '@vitepress-code-preview/container/dist/style.css'
 import 'virtual:uno.css'